Move some functions to common
